requirements.txt
pyautogen==0.3.0
pyautogen[jupyter-executor]
pillow
joblib
matplotlib
numpy
networkx
scipy
datasets
azure==4.0.0
azure-identity
ms-swift
qwen-vl-utils==0.0.11
torch
torchvision
tqdm
anls
scikit-learn
vllm==0.10.2
flaml[automl]
bm3d
opencv-contrib-python
loguru
